Parse files in parallel and only parse buildable files
This commit introduces two changes.  The first is to parse other package
files in parallel.  The second is to only parse files that would be
considered buildable by go/build.

Using go/build.Import() is relatively slow and cannot be parallelized
(it must run before the parallel parse).  To get around using go/build,
this commit adds a package 'buildutil', which is a slimmed down version
of github.com/charlievieth/buildutil, that can quickly determine if a
package should be included.  In order for 'buildutil' to respect the
PackedContext passed in by the client - it was added as a field to
suggest.Config, which is populated by the server.