Cache imported packages
This commit adds a types.Importer that caches packages.  The uses the
filepath of the object/archive file associated with a package as the key
so that vendoring is supported (it also makes the cache relatively simple).
The timestamp of the object/archive file is used to determine when a
package should be re-imported.